Update JS and styles
