+++ www/public/app/css/components/form-type-selector.css
@@ -1,19 +1,32 @@
 .form-type-selector {
     position: relative;
-    padding: 16px 25px 0;
+    padding: 0 var(--islandPadding);
     font-size: 13px;
     color: var(--color-gray);
-    background-color: var(--lightBlueBackground);
 
     @media (--tablet) {
-        padding: 16px 45px 0 25px;
+        padding: 0 35px 0 var(--islandPadding);
     }
+
+    &__separator {
+         position: absolute;
+         left: 0;
+         right: 0;
+         bottom: 0;
+         height: 1px;
+         background-color: var(--lightGrayLine);
+     }
 
     &__content {
          display: flex;
-         max-width: 590px;
+         max-width: var(--editor-max-width);
          overflow-x: auto;
+         overflow-y: hidden;
          margin: 0 auto;
+
+         @media (--tablet) {
+             margin: 0;
+         }
      }
 
     &__item {
@@ -22,20 +35,16 @@
          position: relative;
          flex-basis: auto;
          margin-right: 20px;
-         padding-bottom: 16px;
+         padding: 16px 0;
          flex-shrink: 0;
+         cursor: pointer;
 
         &:last-of-type {
              margin-right: 0;
         }
 
         &:hover {
-             color: var(--redColor);
-             cursor: pointer;
-         }
-
-        label:hover {
-            cursor: pointer;
+             color: var(--color-hover);
         }
 
         input {
@@ -43,7 +52,7 @@
         }
 
         svg {
-            margin-right: 5px;
+            margin-right: 8px;
         }
 
          &-photo {
@@ -54,11 +63,23 @@
               margin-right: 5px;
          }
 
-        &-blog {
-             order: -1;
-        }
+         &-blog {
+              order: -1;
+         }
 
-        &--selected {
+         &--selected {
+
+            &:after {
+                 bottom: -8px;
+                 left: 50%;
+                 border: 1px solid var(--lightGrayLine);
+                 content: '';
+                 height: 12px;
+                 width: 12px;
+                 background-color: #fff;
+                 position: absolute;
+                 transform: translateX(-50%) rotate(45deg);
+            }
 
             svg {
                 color: var(--redColor);
@@ -66,29 +87,14 @@
 
             label {
                 color: var(--redColor);
-
-                &:after {
-                     bottom: 0;
-                     left: 50%;
-                     border: solid transparent;
-                     content: " ";
-                     height: 0;
-                     width: 0;
-                     position: absolute;
-                     border-bottom-color: #fff;
-                     border-width: 15px;
-                     margin-left: -15px;
-                 }
             }
-        }
+         }
     }
 
     .island-settings {
         position: absolute;
         top: 16px;
-        right: 0;
-        background-color: var(--lightBlueBackground);
-        padding: 0 16px 0 8px;
+        right: 16px;
 
         &:hover {
              color: #4d4d4d;
